Moving subscription classes to their own files
