+++ demos/ASP.Net MVC/Divergent.ITOps.ViewModelComposition/Divergent.ITOps.ViewModelComposition.csproj
@@ -88,6 +88,8 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RequestContext.cs" />
     <Compile Include="StringExtensions.cs" />
+    <Compile Include="Subscription %28generic%29.cs" />
+    <Compile Include="Subscription.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
