+++ demos/ASP.Net MVC/Divergent.ITOps.ViewModelComposition/Subscription (generic).cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace Divergent.ITOps.ViewModelComposition
+{
+    class Subscription<T> : Subscription
+    {
+        private Func<dynamic, T, RequestContext, Task> subscription;
+
+        public Subscription(Func<dynamic, T, RequestContext, Task> subscription)
+        {
+            this.subscription = subscription;
+        }
+
+        public override Task Invoke(dynamic viewModel, object @event, RequestContext requestContext)
+        {
+            return subscription(viewModel, (T)@event, requestContext);
+        }
+    }
+}
