+++ demos/ASP.Net MVC/Divergent.ITOps.ViewModelComposition/Subscription.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Divergent.ITOps.ViewModelComposition
+{
+    abstract class Subscription
+    {
+        public abstract Task Invoke(dynamic viewModel, object @event, RequestContext requestContext);
+    }
+}
